requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
pyyaml>=5.4.0
jinja2>=3.0.0
yfinance>=0.2.0
QuantLib>=1.42.0

# Agent layer (Phase 1+)
google-genai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
sse-starlette>=2.0.0
pydantic>=2.0.0
